# server/src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlexCharge Subscription Engine API",
    "description": "API documentation for the FlexCharge Subscription Engine. Supports Tenant Identity, API Key Authentication, Plan and Customer management, Webhook configs, and Customer Portal verification.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:7000/api",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/plans": {
      "get": {
        "summary": "List Plans",
        "description": "Retrieve all plans belonging to this tenant.",
        "tags": ["Plans"],
        "security": [{ "apiKeyAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of plans" }
        }
      },
      "post": {
        "summary": "Create Plan",
        "description": "Create a new subscription plan. Plan amounts should be stored in Kobo (e.g. ₦1,000 = 100000).",
        "tags": ["Plans"],
        "security": [{ "apiKeyAuth": [] }, { "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "amount", "interval"],
                "properties": {
                  "name": { "type": "string", "example": "Pro Monthly Plan" },
                  "amount": { "type": "integer", "example": 500000 },
                  "currency": { "type": "string", "default": "NGN", "example": "NGN" },
                  "interval": { "type": "string", "enum": ["weekly", "monthly", "quarterly", "yearly"], "example": "monthly" },
                  "trialDays": { "type": "integer", "default": 0, "example": 7 },
                  "features": { "type": "array", "items": { "type": "string" }, "example": ["Uncapped features", "Priority Support"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Plan created successfully" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide the Tenant JWT token returned on Registration/Login."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Provide the programmatic API key issued to the Tenant."
      },
      "portalAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide the short-lived Portal token returned by POST /portal/sessions."
      }
    }
  }
}